Rename main to cli.
